pyproject.toml
[tool.poetry]
name = "login-with-flask"
version = "0.1.0"
description = ""
authors = ["Pyunghyuk Yoo <62975595+purple-lion@users.noreply.github.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-Login = "^0.5.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
